backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.60",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-phonenumber-field>=8.4.0",
    "django-rest-framework>=0.1.0",
    "django-storages>=1.14.6",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.15",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "storage>=0.0.4.3",
]
